Add files via upload
hi guys! I am sending the EDA for the 2nd table so Filippa can start and see what is going on and also you if you have any suggestions etc. I tried my best it took me a lot of time and that is why I didn't do the 3rd, but will tomorrow asap!